fastapi_blog/pyproject.toml
[project]
name = "fastapi-blog"
version = "0.1.0"
description = ""
authors = [
    {name = "Tomáš Štucka",email = "Tomas00217@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi[standard] (>=0.115.11,<0.116.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "slugify (>=0.0.1,<0.0.2)",
    "alembic (>=1.15.1,<2.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "djlint (>=1.36.4,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "bleach (>=6.2.0,<7.0.0)",
    "cloudinary (>=1.42.2,<2.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "fastapi-login (>=1.10.3,<2.0.0)",
    "itsdangerous (>=2.2.0,<3.0.0)",
    "wtforms (>=3.2.1,<4.0.0)",
    "starlette-wtf (>=0.4.5,<0.5.0)",
    "starlette-admin (>=0.14.1,<0.15.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "anyio (>=4.9.0,<5.0.0)",
    "pytest-asyncio (>=0.25.3,<0.26.0)",
    "httpx (>=0.28.1,<0.29.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
